uk-esos-app-web/src/app/shared/pipes/legal-entity-type.pipe.ts
import { Pipe, PipeTransform } from '@angular/core';

import { LegalEntityDTO } from 'esos-api';

@Pipe({ name: 'legalEntityType' })
export class LegalEntityTypePipe implements PipeTransform {
  transform(status?: LegalEntityDTO['type']): string {
    switch (status) {
      case 'LIMITED_COMPANY':
        return 'Limited Company';
      case 'OTHER':
        return 'Other';
      case 'PARTNERSHIP':
        return 'Partnership';
      case 'SOLE_TRADER':
        return 'Sole trader';
      default:
        return null;
    }
  }
}
